Move statistics.py outside of core
