--- esmlab/core/__init__.py
@@ -1,2 +1,0 @@
-from __future__ import division, absolute_import, print_function
-from . import statistics
